Ultimos cambios
Cambios minimos para entregar

+++ Libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: dlerma-c <dlerma-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/01 16:53:29 by dlerma-c          #+#    #+#             */
-/*   Updated: 2021/08/11 13:31:04 by dlerma-c         ###   ########.fr       */
+/*   Updated: 2021/08/22 14:51:22 by dlerma-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	const char		*str;
-	unsigned int	i;
+	const char	*str;
+	size_t		i;
 
 	str = s;
 	i = 0;
